requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==23.2.1
